+++ test/libsolidity/memoryGuardTests/unmarked_with_memory_access_mcopy.sol
@@ -1,0 +1,18 @@
+contract C {
+    constructor() {
+        assembly {
+            mcopy(1000, 2000, 1000)
+        }
+    }
+
+    function copy() public {
+        assembly {
+            mcopy(1000, 2000, 1000)
+        }
+    }
+}
+// ====
+// EVMVersion: >=cancun
+// ----
+// :C(creation) false
+// :C(runtime) false