+++ test/tools/yulInterpreter/EVMInstructionInterpreter.h
@@ -56,6 +56,19 @@ void copyZeroExtended(
 	size_t _size
 );
 
+/// Copy @a _size bytes of @a _source at offset @a _sourceOffset to
+/// @a _target at offset @a _targetOffset. Behaves as if @a _source would
+/// continue with an infinite sequence of zero bytes beyond its end.
+/// When target and source areas overlap, behaves as if the data was copied
+/// using an intermediate buffer.
+void copyZeroExtendedWithOverlap(
+	std::map<u256, uint8_t>& _target,
+	std::map<u256, uint8_t> const& _source,
+	size_t _targetOffset,
+	size_t _sourceOffset,
+	size_t _size
+);
+
 struct InterpreterState;
 
 /**